+++ src/main/kotlin/net/henbit/comp/actions/RunNewCommandAction.kt
@@ -1,11 +1,19 @@
 package net.henbit.comp.actions
 
+import cc.ekblad.toml.decode
+import cc.ekblad.toml.model.TomlDocument
+import cc.ekblad.toml.serialization.CollectionSyntax
+import cc.ekblad.toml.serialization.InlineListMode
+import cc.ekblad.toml.tomlMapper
+import cc.ekblad.toml.tomlSerializer
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.diagnostic.Logger
+import com.intellij.util.io.exists
 import io.ktor.utils.io.errors.*
 import net.henbit.comp.Utils
 import net.henbit.comp.dialog.RunNewCommandDialog
+import kotlin.io.path.Path
 
 
 class RunNewCommandAction : AnAction() {
@@ -13,6 +21,10 @@ class RunNewCommandAction : AnAction() {
     companion object {
         val LOGGER = Logger.getInstance(this::class.java)
         const val TAB_NAME = "ComP"
+    }
+
+    data class Cargo(val workspace: Workspace) {
+        data class Workspace(val members: List<String>)
     }
 
     @Suppress("DuplicatedCode")
@@ -33,6 +45,23 @@ class RunNewCommandAction : AnAction() {
                     widget.executeCommand(
                         "cargo compete new $contestId"
                     )
+                    val projectBasePath = project.basePath?.let { Path(it) }
+                    val cargoFilePath = projectBasePath?.resolve("cargo.toml")
+                    if (cargoFilePath != null) {
+                        if (cargoFilePath.exists()) {
+                            val mapper = tomlMapper { }
+                            val cargo = mapper.decode<Cargo>(cargoFilePath)
+                            val mutableMembers = cargo.workspace.members.toMutableList()
+                            mutableMembers.add(contestId)
+                            val newCargo = Cargo(Cargo.Workspace(mutableMembers.toList()))
+                            val serializer = tomlSerializer {
+                                inlineListMode(InlineListMode.MultiLine)
+                                preferListSyntax(CollectionSyntax.Inline)
+                            }
+                            val toml = mapper.encode(newCargo) as TomlDocument
+                            serializer.write(toml, cargoFilePath.toFile().outputStream())
+                        }
+                    }
 
                 } catch (e: IOException) {
                     LOGGER.error("Cannot execute command in local terminal. Error:$e")